convert widget
